+++ lib/spack/spack/build_environment.py
@@ -881,7 +881,7 @@ def fork(pkg, function, dirty, fake, context='build', **kwargs):
 
             # build up some context from the offending package so we can
             # show that, too.
-            if exc_type is not spack.package.TestFailure:
+            if exc_type is not spack.install_test.TestFailure:
                 package_context = get_package_context(traceback.extract_tb(tb))
             else:
                 package_context = []
@@ -893,9 +893,7 @@ def fork(pkg, function, dirty, fake, context='build', **kwargs):
             test_log = None
             if context == 'test':
                 test_log = os.path.join(
-                    spack.util.path.canonicalize_path(
-                        spack.config.get('config:test_stage')),
-                    kwargs.get('test_name'),
+                    pkg.test_suite.stage,
                     spack.install_test.TestSuite.test_log_name(pkg.spec))
 
             # make a pickleable exception to send to parent.
